Refactor pagination trait placement
