+++ src/api/v1/handlers/groups.rs
@@ -3,9 +3,9 @@ use crate::db::DbPool;
 use crate::errors::ApiError;
 use crate::extractors::{AdminAccess, UserAccess};
 use crate::models::group::{GroupID, NewGroup, UpdateGroup};
-use crate::models::pagination::prepare_db_pagination;
 use crate::models::search::parse_query_parameter;
 use crate::models::{Group, User, UserID};
+use crate::pagination::prepare_db_pagination;
 use crate::utilities::response::{json_response, json_response_created, paginated_json_response};
 use actix_web::{delete, get, http::StatusCode, patch, post, routes, web, HttpRequest, Responder};
 use serde::{Deserialize, Serialize};
