+++ src/api/v1/handlers/relations.rs
@@ -2,12 +2,12 @@ use crate::api::openapi::ApiErrorResponse;
 use crate::db::DbPool;
 use crate::errors::ApiError;
 use crate::extractors::UserAccess;
-use crate::models::pagination::prepare_db_pagination;
 use crate::models::search::parse_query_parameter;
 use crate::models::{
     HubuumClassRelation, HubuumClassRelationID, HubuumObjectRelation, HubuumObjectRelationID,
     NamespaceID, NewHubuumClassRelation, NewHubuumObjectRelation, Permissions,
 };
+use crate::pagination::prepare_db_pagination;
 
 use crate::can;
 use crate::db::traits::UserPermissions;
