+++ src/api/v1/handlers/users.rs
@@ -2,10 +2,10 @@ use crate::api::openapi::ApiErrorResponse;
 use crate::db::DbPool;
 use crate::errors::ApiError;
 use crate::extractors::{AdminAccess, AdminOrSelfAccess, UserAccess};
-use crate::models::pagination::prepare_db_pagination;
 use crate::models::search::parse_query_parameter;
 use crate::models::user::{NewUser, UpdateUser, UserID};
 use crate::models::{Group, User, UserToken};
+use crate::pagination::prepare_db_pagination;
 use crate::utilities::response::{json_response, json_response_created, paginated_json_response};
 use actix_web::{delete, get, http::StatusCode, patch, routes, web, HttpRequest, Responder};
 use serde_json::json;
