+++ src/config.rs
@@ -10,6 +10,9 @@ use serde::{Deserialize, Serialize};
 
 use crate::errors::ApiError;
 
+pub const DEFAULT_PAGE_LIMIT: usize = 100;
+pub const MAX_PAGE_LIMIT: usize = 250;
+
 #[derive(ValueEnum, Debug, Deserialize, Serialize, Clone, Copy, PartialEq, Eq)]
 #[serde(rename_all = "lowercase")]
 pub enum TlsBackend {
@@ -61,6 +64,22 @@ pub struct AppConfig {
     #[clap(long, env = "HUBUUM_DB_POOL_SIZE", default_value_t = 10)]
     pub db_pool_size: u32,
 
+    /// Default number of items returned by cursor-paginated list endpoints
+    #[clap(
+        long,
+        env = "HUBUUM_DEFAULT_PAGE_LIMIT",
+        default_value_t = DEFAULT_PAGE_LIMIT
+    )]
+    pub default_page_limit: usize,
+
+    /// Maximum number of items allowed for cursor-paginated list endpoints
+    #[clap(
+        long,
+        env = "HUBUUM_MAX_PAGE_LIMIT",
+        default_value_t = MAX_PAGE_LIMIT
+    )]
+    pub max_page_limit: usize,
+
     /// The name of the admin group
     #[clap(long, env = "HUBUUM_ADMIN_GROUPNAME", default_value = "admin")]
     pub admin_groupname: String,
@@ -88,8 +107,43 @@ pub struct AppConfig {
     pub tls_backend: Option<TlsBackend>,
 }
 
+impl AppConfig {
+    fn validate(self) -> Result<Self, ApiError> {
+        if self.default_page_limit == 0 {
+            return Err(ApiError::BadRequest(
+                "default_page_limit must be greater than 0".to_string(),
+            ));
+        }
+
+        if self.max_page_limit == 0 {
+            return Err(ApiError::BadRequest(
+                "max_page_limit must be greater than 0".to_string(),
+            ));
+        }
+
+        if self.default_page_limit > self.max_page_limit {
+            return Err(ApiError::BadRequest(format!(
+                "default_page_limit ({}) must be less than or equal to max_page_limit ({})",
+                self.default_page_limit, self.max_page_limit
+            )));
+        }
+
+        Ok(self)
+    }
+}
+
 #[cfg(not(test))]
-pub static CONFIG: Lazy<RwLock<AppConfig>> = Lazy::new(|| RwLock::new(AppConfig::parse()));
+fn load_config() -> Result<AppConfig, ApiError> {
+    AppConfig::try_parse()
+        .map_err(|e| ApiError::BadRequest(format!("Invalid configuration: {e}")))?
+        .validate()
+}
+
+#[cfg(not(test))]
+pub static CONFIG: Lazy<RwLock<AppConfig>> = Lazy::new(|| {
+    let config = load_config().unwrap_or_else(|e| panic!("Invalid application configuration: {e}"));
+    RwLock::new(config)
+});
 
 #[cfg(not(test))]
 pub fn get_config() -> Result<RwLockReadGuard<'static, AppConfig>, ApiError> {
@@ -121,7 +175,7 @@ fn get_config_from_env() -> Result<AppConfig, ApiError> {
         })
     }
 
-    Ok(AppConfig {
+    let config = AppConfig {
         bind_ip: env_or_default("HUBUUM_BIND_IP", "127.0.0.1"),
         port: env_or_default("HUBUUM_BIND_PORT", "8080")
             .parse()
@@ -134,12 +188,20 @@ fn get_config_from_env() -> Result<AppConfig, ApiError> {
         db_pool_size: env_or_default("HUBUUM_DB_POOL_SIZE", "2")
             .parse()
             .unwrap_or(5),
+        default_page_limit: env_or_default("HUBUUM_DEFAULT_PAGE_LIMIT", "100")
+            .parse()
+            .unwrap_or(DEFAULT_PAGE_LIMIT),
+        max_page_limit: env_or_default("HUBUUM_MAX_PAGE_LIMIT", "250")
+            .parse()
+            .unwrap_or(MAX_PAGE_LIMIT),
         admin_groupname: env_or_default("HUBUUM_ADMIN_GROUPNAME", "admin"),
         tls_cert_path: env_or_default_opt("HUBUUM_TLS_CERT_PATH", None),
         tls_key_path: env_or_default_opt("HUBUUM_TLS_KEY_PATH", None),
         tls_key_passphrase: env_or_default_opt("HUBUUM_TLS_KEY_PASSPHRASE", None),
         tls_backend: env_or_default_tls_backend("HUBUUM_TLS_BACKEND"),
-    })
+    };
+
+    config.validate()
 }
 
 #[cfg(test)]
@@ -154,7 +216,10 @@ mod tests {
 
     use clap::Parser;
 
-    use super::{get_config_from_env, AppConfig, TlsBackend, TEST_ENV_LOCK};
+    use super::{
+        get_config_from_env, AppConfig, TlsBackend, DEFAULT_PAGE_LIMIT, MAX_PAGE_LIMIT,
+        TEST_ENV_LOCK,
+    };
 
     struct EnvVarGuard {
         key: &'static str,
@@ -218,5 +283,49 @@ mod tests {
         assert!(error.to_string().contains("bogus"));
         assert!(error.to_string().contains("rustls"));
         assert!(error.to_string().contains("openssl"));
+    }
+
+    #[test]
+    fn page_limits_are_parsed_from_env() {
+        let _lock = TEST_ENV_LOCK.lock().unwrap();
+        let _default_guard = EnvVarGuard::set("HUBUUM_DEFAULT_PAGE_LIMIT", Some("25"));
+        let _max_guard = EnvVarGuard::set("HUBUUM_MAX_PAGE_LIMIT", Some("75"));
+
+        let parsed = AppConfig::try_parse_from(["hubuum-server"]).unwrap();
+        let loaded = get_config_from_env().unwrap();
+
+        assert_eq!(parsed.default_page_limit, 25);
+        assert_eq!(parsed.max_page_limit, 75);
+        assert_eq!(loaded.default_page_limit, 25);
+        assert_eq!(loaded.max_page_limit, 75);
+    }
+
+    #[test]
+    fn page_limits_default_when_env_vars_are_unset() {
+        let _lock = TEST_ENV_LOCK.lock().unwrap();
+        let _default_guard = EnvVarGuard::set("HUBUUM_DEFAULT_PAGE_LIMIT", None);
+        let _max_guard = EnvVarGuard::set("HUBUUM_MAX_PAGE_LIMIT", None);
+
+        let parsed = AppConfig::try_parse_from(["hubuum-server"]).unwrap();
+        let loaded = get_config_from_env().unwrap();
+
+        assert_eq!(parsed.default_page_limit, DEFAULT_PAGE_LIMIT);
+        assert_eq!(parsed.max_page_limit, MAX_PAGE_LIMIT);
+        assert_eq!(loaded.default_page_limit, DEFAULT_PAGE_LIMIT);
+        assert_eq!(loaded.max_page_limit, MAX_PAGE_LIMIT);
+    }
+
+    #[test]
+    fn page_limits_are_validated() {
+        let _lock = TEST_ENV_LOCK.lock().unwrap();
+        let _default_guard = EnvVarGuard::set("HUBUUM_DEFAULT_PAGE_LIMIT", Some("80"));
+        let _max_guard = EnvVarGuard::set("HUBUUM_MAX_PAGE_LIMIT", Some("40"));
+
+        let error = get_config_from_env().unwrap_err();
+
+        assert_eq!(
+            error.to_string(),
+            "default_page_limit (80) must be less than or equal to max_page_limit (40)"
+        );
     }
 }
