+++ src/models/mod.rs
@@ -5,7 +5,6 @@ pub mod group;
 pub mod namespace;
 pub mod object;
 pub mod output;
-pub mod pagination;
 pub mod permissions;
 pub mod relation;
 pub mod search;
@@ -20,7 +19,6 @@ pub use crate::models::group::*;
 pub use crate::models::namespace::*;
 pub use crate::models::object::*;
 pub use crate::models::output::*;
-pub use crate::models::pagination::*;
 pub use crate::models::permissions::*;
 pub use crate::models::relation::*;
 pub use crate::models::token::*;
