+++ src/models/traits/output.rs
@@ -4,11 +4,14 @@ use tracing::warn;
 use crate::db::DbPool;
 use crate::errors::ApiError;
 use crate::models::group::Group;
+use crate::models::search::{FilterField, SortParam};
 use crate::models::{
     GroupPermission, HubuumClass, HubuumClassExpanded, Namespace, NamespaceID, Permission,
     Permissions, PermissionsList,
 };
-use crate::traits::SelfAccessors;
+use crate::traits::{
+    CursorPaginated, CursorSqlField, CursorSqlMapping, CursorSqlType, CursorValue, SelfAccessors,
+};
 
 pub trait FromTuple<T> {
     fn from_tuple(t: (Group, T)) -> Self;
@@ -93,5 +96,160 @@ impl ExpandNamespaceFromMap<HubuumClassExpanded> for HubuumClass {
             created_at: self.created_at,
             updated_at: self.updated_at,
         }
+    }
+}
+
+impl CursorPaginated for HubuumClassExpanded {
+    fn supports_sort(field: &FilterField) -> bool {
+        matches!(
+            field,
+            FilterField::Id
+                | FilterField::Name
+                | FilterField::Namespaces
+                | FilterField::NamespaceId
+                | FilterField::CreatedAt
+                | FilterField::UpdatedAt
+        )
+    }
+
+    fn cursor_value(&self, field: &FilterField) -> Result<CursorValue, ApiError> {
+        Ok(match field {
+            FilterField::Id => CursorValue::Integer(self.id as i64),
+            FilterField::Name => CursorValue::String(self.name.clone()),
+            FilterField::Namespaces | FilterField::NamespaceId => {
+                CursorValue::Integer(self.namespace.id as i64)
+            }
+            FilterField::CreatedAt => CursorValue::DateTime(self.created_at),
+            FilterField::UpdatedAt => CursorValue::DateTime(self.updated_at),
+            _ => {
+                return Err(ApiError::BadRequest(format!(
+                    "Field '{}' is not orderable for classes",
+                    field
+                )))
+            }
+        })
+    }
+
+    fn default_sort() -> Vec<SortParam> {
+        vec![SortParam {
+            field: FilterField::Id,
+            descending: false,
+        }]
+    }
+
+    fn tie_breaker_sort() -> Vec<SortParam> {
+        Self::default_sort()
+    }
+}
+
+impl CursorSqlMapping for HubuumClassExpanded {
+    fn sql_field(field: &FilterField) -> Result<CursorSqlField, ApiError> {
+        Ok(match field {
+            FilterField::Id => CursorSqlField {
+                column: "hubuumclass.id",
+                sql_type: CursorSqlType::Integer,
+                nullable: false,
+            },
+            FilterField::Name => CursorSqlField {
+                column: "hubuumclass.name",
+                sql_type: CursorSqlType::String,
+                nullable: false,
+            },
+            FilterField::Namespaces | FilterField::NamespaceId => CursorSqlField {
+                column: "hubuumclass.namespace_id",
+                sql_type: CursorSqlType::Integer,
+                nullable: false,
+            },
+            FilterField::CreatedAt => CursorSqlField {
+                column: "hubuumclass.created_at",
+                sql_type: CursorSqlType::DateTime,
+                nullable: false,
+            },
+            FilterField::UpdatedAt => CursorSqlField {
+                column: "hubuumclass.updated_at",
+                sql_type: CursorSqlType::DateTime,
+                nullable: false,
+            },
+            _ => {
+                return Err(ApiError::BadRequest(format!(
+                    "Field '{}' is not orderable for classes",
+                    field
+                )))
+            }
+        })
+    }
+}
+
+impl CursorPaginated for GroupPermission {
+    fn supports_sort(field: &FilterField) -> bool {
+        matches!(
+            field,
+            FilterField::Id
+                | FilterField::Name
+                | FilterField::Groupname
+                | FilterField::CreatedAt
+                | FilterField::UpdatedAt
+        )
+    }
+
+    fn cursor_value(&self, field: &FilterField) -> Result<CursorValue, ApiError> {
+        Ok(match field {
+            FilterField::Id => CursorValue::Integer(self.permission.id as i64),
+            FilterField::Name | FilterField::Groupname => {
+                CursorValue::String(self.group.groupname.clone())
+            }
+            FilterField::CreatedAt => CursorValue::DateTime(self.permission.created_at),
+            FilterField::UpdatedAt => CursorValue::DateTime(self.permission.updated_at),
+            _ => {
+                return Err(ApiError::BadRequest(format!(
+                    "Field '{}' is not orderable for group permissions",
+                    field
+                )))
+            }
+        })
+    }
+
+    fn default_sort() -> Vec<SortParam> {
+        vec![SortParam {
+            field: FilterField::Id,
+            descending: false,
+        }]
+    }
+
+    fn tie_breaker_sort() -> Vec<SortParam> {
+        Self::default_sort()
+    }
+}
+
+impl CursorSqlMapping for GroupPermission {
+    fn sql_field(field: &FilterField) -> Result<CursorSqlField, ApiError> {
+        Ok(match field {
+            FilterField::Id => CursorSqlField {
+                column: "permissions.id",
+                sql_type: CursorSqlType::Integer,
+                nullable: false,
+            },
+            FilterField::Name | FilterField::Groupname => CursorSqlField {
+                column: "groups.groupname",
+                sql_type: CursorSqlType::String,
+                nullable: false,
+            },
+            FilterField::CreatedAt => CursorSqlField {
+                column: "permissions.created_at",
+                sql_type: CursorSqlType::DateTime,
+                nullable: false,
+            },
+            FilterField::UpdatedAt => CursorSqlField {
+                column: "permissions.updated_at",
+                sql_type: CursorSqlType::DateTime,
+                nullable: false,
+            },
+            _ => {
+                return Err(ApiError::BadRequest(format!(
+                    "Field '{}' is not orderable for group permissions",
+                    field
+                )))
+            }
+        })
     }
 }
